move to_color_strings to plotting._utils
